Reorganize classes, initial Gtk4 support
