+++ test/qz/utils/gtk/GtkTests.java
@@ -1,13 +1,12 @@
-package qz.utils;
+package qz.utils.gtk;
 
 import com.github.zafarkhaja.semver.Version;
 import org.testng.Assert;
 import org.testng.SkipException;
 import org.testng.annotations.Test;
+import qz.utils.SystemUtilities;
 
-import static qz.utils.GtkUtilities.*;
-
-public class GtkUtilitiesTests {
+public class GtkTests {
     @Test
     public void testGtkAvailability() {
         switch(SystemUtilities.getOs()) {
@@ -15,7 +14,7 @@ public class GtkUtilitiesTests {
             case MAC:
                 throw new SkipException("GTK is not needed on macOS or Windows");
         }
-        GTK GTK_INSTANCE = GTK.getInstance(false);
+        Gtk GTK_INSTANCE = Gtk.getInstance(false);
         Assert.assertNotNull(GTK_INSTANCE);
         Assert.assertNotEquals(GTK_INSTANCE.getVersion(), Version.of(0));
     }
